secu type route

+++ src/Controller/TypeController.php
@@ -30,6 +30,9 @@ class TypeController extends AbstractController
     #[Route('/new', name: 'app_type_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
         $type = new Type();
         $form = $this->createForm(TypeType::class, $type);
         $form->handleRequest($request);
@@ -50,6 +53,9 @@ class TypeController extends AbstractController
     #[Route('/{id}', name: 'app_type_show', methods: ['GET'])]
     public function show(Type $type): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
         return $this->render('type/show.html.twig', [
             'type' => $type,
         ]);
@@ -58,6 +64,10 @@ class TypeController extends AbstractController
     #[Route('/{id}/edit', name: 'app_type_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Type $type, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
+
         $form = $this->createForm(TypeType::class, $type);
         $form->handleRequest($request);
 
@@ -76,6 +86,9 @@ class TypeController extends AbstractController
     #[Route('/{id}', name: 'app_type_delete', methods: ['POST'])]
     public function delete(Request $request, Type $type, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$type->getId(), $request->getPayload()->get('_token'))) {
             $entityManager->remove($type);
             $entityManager->flush();
